\name{random_hour}
\alias{random_hour}

\title{
Returns a random hour in 0,...,11.
}
\description{
It uses the normal distribution with mean 6 to generate random values. 
}
\usage{
random_hour(n)
}

\arguments{
  \item{n}{
The amount of random hours to be generated.
}
}
\value{
A vector containing n hours.
}
\author{
Thiago Sousa
}
